+++ scopus/tests/test_AuthorRetrieval.py
@@ -31,15 +31,15 @@ def test_cited_by_count():
     assert_true(int(au.cited_by_count) >= 6066)
 
 
-def test_coauthor_count():
-    assert_true(int(au.coauthor_count) >= 164)
-
-
 def test_classificationgroup():
     groups = au.classificationgroup
     assert_true(isinstance(groups, list))
     assert_true(len(groups) > 0)
     assert_true(('1906', '1') in groups)  # frequency might differ
+
+
+def test_coauthor_count():
+    assert_true(int(au.coauthor_count) >= 164)
 
 
 def test_coauthor_link():
@@ -77,6 +77,12 @@ def test_h_index():
     assert_true(int(au.h_index) >= 27)
 
 
+def test_historical_identifier():
+    expected = ['35787230500', '36488127000', '54974425600', '55004143700',
+                '55004143800', '57057263700', '56641032000', '36747787600']
+    assert_equal(au.historical_identifier, expected)
+
+
 def test_identifier():
     assert_equal(au.identifier, "7004212771")
 
@@ -97,10 +103,6 @@ def test_journal_history():
     assert_true(expected in hist)
 
 
-def test_orcid():
-    assert_equal(au.orcid, '0000-0003-2625-9232')
-
-
 def test_name_variants():
     names = au.name_variants
     assert_true(isinstance(names, list))
@@ -109,19 +111,18 @@ def test_name_variants():
     assert_equal(str(type(names[0])), expected)
 
 
-def test_surname():
-    assert_equal(au.surname, 'Kitchin')
+def test_orcid():
+    assert_equal(au.orcid, '0000-0003-2625-9232')
+
+
+def test_publication_range():
+    assert_equal(au.publication_range[0], '1995')
+    assert_true(int(au.publication_range[1]) >= 2018)
 
 
 def test_scopus_author_link():
     expected = 'http://api.elsevier.com/content/author/author_id/7004212771'
     assert_equal(au.scopus_author_link, expected)
-
-
-def test_self_link():
-    expected = 'https://www.scopus.com/authid/detail.uri?partnerID=HzOxMe3b&'\
-               'authorId=7004212771&origin=inward'
-    assert_equal(au.self_link, expected)
 
 
 def test_search_link():
@@ -130,9 +131,14 @@ def test_search_link():
     assert_equal(au.search_link, expected)
 
 
-def test_publication_range():
-    assert_equal(au.publication_range[0], '1995')
-    assert_true(int(au.publication_range[1]) >= 2018)
+def test_self_link():
+    expected = 'https://www.scopus.com/authid/detail.uri?partnerID=HzOxMe3b&'\
+               'authorId=7004212771&origin=inward'
+    assert_equal(au.self_link, expected)
+
+
+def test_status():
+    assert_equal(au.status, "update")
 
 
 def test_subject_areas():
@@ -141,6 +147,10 @@ def test_subject_areas():
     assert_true(len(areas) > 0)
     expected = "<class 'scopus.author_retrieval.Subjectarea'>"
     assert_equal(str(type(areas[0])), expected)
+
+
+def test_surname():
+    assert_equal(au.surname, 'Kitchin')
 
 
 def test_url():